Fix scroll styling
